# cmd/payment/docs/swagger.yaml
basePath: /
definitions:
...
  model.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.createUserDTO:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para el manejo de pagos
  title: API de Pagos
  version: "1"
paths:
...
  /api/user/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un usuario por su ID.
      parameters:
      - description: ID del Usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Usuario eliminado correctamente
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Elimina un usuario
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Obtiene un usuario por su ID, excluyendo la contraseña en la respuesta.
      parameters:
      - description: ID del Usuario
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuario encontrado
          schema:
            $ref: '#/definitions/model.User'
        "401":
          description: No autorizado
          schema:
            type: string
        "404":
          description: Usuario no encontrado
          schema:
            type: string
        "500":
          description: Error al obtener el usuario
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Obtiene un usuario
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Actualiza la información de un usuario existente por su ID.
      parameters:
      - description: ID del Usuario
        in: path
        name: id
        required: true
        type: integer
      - description: Información actualizada del Usuario
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Error en la solicitud
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Usuario no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualiza un usuario
      tags:
      - user
...
  /user:
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario con username y password proporcionados.
      parameters:
      - description: Credenciales de usuario
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/model.createUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado con ID y username
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error al decodificar solicitud o usuario/password faltantes
          schema:
            type: string
        "409":
          description: Nombre de usuario ya existe
          schema:
            type: string
        "500":
          description: Error al hashear la contraseña / Error al crear el usuario
          schema:
            type: string
      summary: Crea un nuevo usuario
      tags:
      - user
swagger: "2.0"